Refactor project structure - one service for unix and windows
