--- src/Domain0.Nancy.Kestrel/Domain0.Nancy.Kestrel.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <DebugType>portable</DebugType>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Hosting" Version="2.2.7" />
-    <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Owin" Version="3.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Domain0.Nancy\Domain0.Nancy.csproj" />
-    <ProjectReference Include="..\Domain0.Repository.FastSql\Domain0.Repository.FastSql.csproj" />
-  </ItemGroup>
-
-</Project>
