--- src/Domain0.Nancy.Kestrel/Properties/launchSettings.json
@@ -1,12 +1,0 @@
-{
-  "profiles": {
-    "Domain0.Nancy.Kestrel": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      },
-      "applicationUrl": "http://localhost:56200/"
-    }
-  }
-}
