+++ src/Domain0.Nancy/Modules/AdminModule.cs
