+++ src/Domain0.Nancy/Modules/ApplicationModule.cs
