+++ src/Domain0.Nancy/Modules/EmailModule.cs
