+++ src/Domain0.Nancy/Modules/LdapModule.cs
