--- src/Domain0.Repository.FastSql/Domain0.Repository.FastSql.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net46;netstandard2.0</TargetFrameworks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Scripts\database.sql" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Domain0.Repository\Domain0.Repository.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Gerakul.FastSql.SqlServer" Version="1.0.0" />
-    <PackageReference Include="Autofac" Version="4.9.4" />
-    <PackageReference Include="NLog" Version="4.6.7" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <PackageReference Include="System.Configuration.ConfigurationManager">
-      <Version>4.6.0</Version>
-    </PackageReference>
-  </ItemGroup>
-
-</Project>
