+++ src/Domain0.Repository/DbManager.cs
