+++ src/Domain0.Repository/FastSql/ApplicationRepository.cs
