+++ src/Domain0.Repository/FastSql/MessageTemplateRepository.cs
