+++ src/Domain0.Repository/Interfaces/IAccessLogRepository.cs
