+++ src/Domain0.Repository/Interfaces/IApplicationRepository.cs
