+++ src/Domain0.Repository/Interfaces/IEmailRequestRepository.cs
