+++ src/Domain0.Repository/Interfaces/IMessageTemplateRepository.cs
