+++ src/Domain0.Repository/Interfaces/IPermissionRepository.cs
