+++ src/Domain0.Repository/Interfaces/IRepository.cs
