+++ src/Domain0.Repository/Interfaces/IRoleRepository.cs
