+++ src/Domain0.Repository/Interfaces/ITokenRegistrationRepository.cs
