+++ src/Domain0.Service/App.config
@@ -4,7 +4,7 @@
     <add name="Database" connectionString="Data Source=.;Initial Catalog=Domain0;Integrated Security=True" />
   </connectionStrings>
   <appSettings>
-    <!-- Base service url if you use https then fill X509 settings -->
+    <!-- Base service url -->
     <add key="Url" value="" />
 
     <!-- Token generation settings -->
