+++ src/Domain0.Service/NLog.config
@@ -1,7 +1,7 @@
 <nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <extensions>
-    <add assembly="Domain0.Nancy.Kestrel" />
+    <add assembly="Domain0.Service" />
   </extensions>
   <targets>
     <target xsi:type="Console" name="console" layout="[${longdate}] ${level}:${message} ${exception:format=tostring}" />
