--- src/Domain0.WinService/App.config
@@ -1,57 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <connectionStrings>
-    <add name="Database" connectionString="Data Source=.;Initial Catalog=Domain0;Integrated Security=True" />
-  </connectionStrings>
-  <appSettings>
-    <!-- Base service url if you use https then fill X509 settings -->
-    <add key="Url" value="" />
-    <!-- File settings for https X509 certficate -->
-    <add key="X509_Filepath" value="" />
-    <add key="X509_Password" value="" />
-    <!-- CertStore settings for https X509 certficate -->
-    <add key="X509_Location" value="" />
-    <add key="X509_StoreName" value="" />
-    <add key="X509_Subject" value="" />
-
-    <!-- Token generation settings -->
-    <add key="Token_Audience" value="*" />
-    <add key="Token_RefreshAudience" value="*" />
-    <add key="Token_Issuer" value="issuer" />
-    <add key="Token_LifeTime" value="15" />
-    <add key="Token_RefreshLifeTime" value="480" />
-    <add key="Token_Secret" value="" />
-    <add key="Token_Private" value="" />
-    <add key="Token_Public" value="" />
-    <add key="Token_Alg" value="HS256" />
-
-    <!-- Email client settings -->
-    <add key="Email_ServerHost" value="" />
-    <add key="Email_Port" value="25" />
-    <add key="Email_Address" value="" />
-    <add key="Email_Username" value="" />
-    <add key="Email_Password" value="" />
-
-    <!-- SqlQueueSmsClientSettings -->
-    <add key="SmsQueueClient_ConnectionString" value="" />
-    <add key="SmsQueueClient_QueueName" value="" />
-
-    <!-- AccountServiceSettings -->
-    <add key="AccountService_MessagesResendCooldown" value="1" />
-    <add key="AccountService_PinExpirationTime" value="15" />
-    <add key="AccountService_EmailCodeExpirationTime" value="120" />
-
-    <!-- SmsGatewaySettings -->
-    <add key="SmsGateway_Host" value="" />
-    <add key="SmsGateway_Token" value="" />
-
-    <!-- Ldap Settings -->
-    <add key="DomainControllerName" value="" />
-    <add key="LdapPort" value="" />
-    <add key="UseSecureSocketLayer" value="" />
-    <add key="LdapProtocolVersion" value="" />
-    <add key="LdapAuthType" value="" />
-    <add key="EmailAttributeName" value="" />
-
-  </appSettings>
-</configuration>
