--- src/Domain0.WinService/Domain0.WinService.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFrameworks>net462</TargetFrameworks>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Update="NLog.config">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Domain0.Nancy\Domain0.Nancy.csproj" />
-    <ProjectReference Include="..\Domain0.Repository.FastSql\Domain0.Repository.FastSql.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Monik.Client.Azure" Version="1.0.24" />
-    <PackageReference Include="Nancy.Hosting.Self" Version="2.0.0" />
-    <PackageReference Include="Topshelf" Version="4.2.1" />
-  </ItemGroup>
-</Project>
