--- src/Domain0.WinService/NLog.config
@@ -1,20 +1,0 @@
-<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <extensions>
-    <add assembly="Domain0.WinService" />
-  </extensions>
-  <targets>
-    <target xsi:type="Console" name="console" layout="[${longdate}] ${level}:${message} ${exception:format=tostring}" />
-    <target xsi:type="File" name="localFile" layout="[${longdate}] ${level}:${message} ${exception:format=tostring}"
-            fileName="${basedir}/logs/logfile.txt" archiveFileName="${basedir}/logs/log.{#}.txt" archiveEvery="Day"
-            archiveNumbering="Rolling" maxArchiveFiles="14" concurrentWrites="true" encoding="utf-8" />
-    <target xsi:type="Monik" name="monik" layout="${message} ${exception:format=tostring}"
-            ConnectionString=""
-            QueueName="" />
-  </targets>
-  <rules>
-    <logger name="*" minlevel="Debug" writeTo="console" />
-    <logger name="*" minlevel="Debug" writeTo="localFile" />
-    <logger name="*" minlevel="Debug" writeTo="monik" />
-  </rules>
-</nlog>
